Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@16049 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNStatusListCtrlHelpers.cpp
@@ -976,7 +976,7 @@ bool CSVNStatusListCtrl::CSorter::operator()
 		{
 			if (result == 0)
 			{
-				result = entry1->lock_date - entry2->lock_date;
+				result = (int)(entry1->lock_date - entry2->lock_date);
 			}
 		}
 	case 12:
